requirements.txt
Flask==3.0.3
requests==2.32.3
curl-cffi==0.7.3
Flask-SQLAlchemy==3.1.1
Bootstrap-Flask==2.4.0
SQLAlchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
WTForms==3.1.2
Flask-WTF==1.2.1
